requirements.txt
Flask~=3.1.0
Flask-Login~=0.6.3
Flask-SQLAlchemy
gunicorn
itsdangerous
Jinja2
MarkupSafe
SQLAlchemy
Werkzeug~=3.1.3
wheel
honcho
Flask-SocketIO~=5.5.1
python-socketio~=5.12.1